+++ src/main/java/com/demoboletto/dto/request/CreateStickerDto.java
@@ -26,6 +26,6 @@ public record CreateStickerDto(
         int rotation,
         @NotNull(message = "scale can not be null")
         @JsonProperty("scale") @Schema(description = "speech scale", example = "0")
-        int scale
+        float scale
 ) {
 }
